Fix pc-parse code for 64bit Windows compilation
Some Linux tweaks as well to ensure everything compiles okay on both
platforms and passes all the regression tests.

+++ pc-parse/ample/putsd.c
@@ -36,7 +36,7 @@ TextControl *		pTextCtl_in;
 {
 char *	p;
 int	c = NUL;
-int	iLength;
+size_t	iLength;
 
 if (bIsCap_in)
 	{
